Move test for received service date set into new test class and delete old outdated class
